str_split

+++ ft_substr.c
@@ -5,8 +5,8 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	size_t			i;
-	size_t			str_len;
+	unsigned int	i;
+	unsigned int	str_len;
 	char			*substr;
 
 	if (!s)
